// frontend/dentistry-front/lib/redux/modules/Clients/clients.interface.ts
export enum BloodSign {
  plus = "plus",
  minus = "minus",
}

export interface IClient {
  id: number; // Унікальний ідентифікатор
  name: string; // Ім'я
  surname: string; // Прізвище
  middle_name?: string; // По батькові (може бути необов'язковим)
  birthdate: string; // Дата народження (як ISO‑рядок для фронта)
  blood_resus: BloodSign; // Резус-фактор (+/-)
  blood_group: number; // Група крові (1–4)
  phone: string; // Телефон
  allergic_diseases?: string; // Алергії (необов'язкове)
  email: string; // Електронна пошта
  isVerified: boolean; // Чи підтверджений акаунт

  created_at?: string; // Дата створення (ISO‑рядок, необов'язкова)
  updated_at?: string; // Дата останнього оновлення (ISO‑рядок, необов'язкова)

  // foreign connections
  //   appointments?: IAppointment[]; // Зв’язки з прийомами, якщо потрібні у фронті
}